requirements.txt
aerich ==0.9.2
aiosqlite==0.22.0
amqp ==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings ==25.1.0
asyncclick ==8.3.0.7
asyncpg==0.31.0
atlastk==0.13.5
bcrypt ==5.0.0
billiard ==4.2.4
celery ==5.6.0
celery-types ==0.24.0
certifi==2025.11.12
cffi ==2.0.0
charset-normalizer ==3.4.4
click==8.3.1
click-didyoumean ==0.3.1
click-plugins==1.1.1.2
click-repl ==0.3.0
colorama ==0.4.6
dictdiffer ==0.9.0
exceptiongroup ==1.3.1
fastapi==0.127.0
filetype ==1.2.0
h11==0.16.0
httpcore ==1.0.9
httpx==0.28.1
idna ==3.11
iso8601==2.1.0
kombu==5.6.1
minio==7.2.20
packaging==25.0
passlib==1.7.4
pip==25.1.1
prompt_toolkit ==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome ==3.23.0
pydantic ==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
PyJWT==2.10.1
pypika-tortoise==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart ==0.0.21
pytz ==2025.2
redis==7.1.0
requests ==2.32.5
ronglian-sms-sdk ==1.0.0
six==1.17.0
starlette==0.50.0
tortoise ==0.1.1
tortoise-orm ==0.25.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata ==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine ==5.1.0
wcwidth==0.2.14
motor==3.6.0
aio-pika==9.4.3
numpy==2.0.2
scikit-learn==1.6.0